incorporate DRBG into build process; address resulting compiler & linker errors & warnings
git-svn-id: https://sparrow.ece.cmu.edu/svn/hyp@1786 1c601a1c-44f4-4d93-90ac-4b632a93592e

+++ trustvisor/trunk/code/app/include/drbg.h
@@ -53,7 +53,7 @@
     switch(0){case 0:case pred:;}
 
 struct INT128_t {
-    int64_t high;
+    long long high;
     uint64_t low;
 } __attribute__ ((packed));
 typedef struct INT128_t INT128;
